[feat/LIG-58] 1. refactor coordinator rpc api 2. remove healthcheck in docker-compose
