+++ proto/common/common.proto
@@ -121,3 +121,26 @@ message Ack {
   // the execution id
   optional ExecutionId execution_id = 6;
 }
+
+
+// Basic information of task
+message TaskInfo {
+  // execution id of task
+  common.ExecutionId execution_id = 1;
+  // information of executors
+  map<uint32, ExecutorInfo> executors_info = 2;
+
+  // Basic information of executor
+  message ExecutorInfo {
+    uint32 executor_id = 1;
+    ExecutorStatus status = 2;
+  }
+
+  // status of executor
+  enum ExecutorStatus {
+    EXECUTOR_STATUS_INITIALIZED = 0;
+    EXECUTOR_STATUS_RUNNING = 1;
+    EXECUTOR_STATUS_TERMINATING = 2;
+    EXECUTOR_STATUS_TERMINATED = 3;
+  }
+}
